fix: keep daemon socket owner check portable (#8)
